Prevent users deleting thread which don't belong to them

+++ app/Http/Controllers/ThreadsController.php
@@ -115,6 +115,13 @@ class ThreadsController extends Controller
      */
     public function destroy(Channel $channel, Thread $thread)
     {
+        if($thread->user_id != auth()->id()){
+            if(request()->wantsJson()) {
+                return response(['status' => 'Permision Denied'], 403);
+            }
+            return redirect('/login');
+        }
+        
         $thread->delete();
 
         if(request()->wantsJson()) {
